newlisp: remove devel
http://www.newlisp.org/downloads/development/inprogress/readme.txt

"Files in this directory are not official development releases, but work
in progress. Source .tgz packages or files may appear or may be changed
unannounced and with unchanged version numbers or CHANGES files."

So having a devel spec in this formula will just break CI every time
upstream bumps the tarball without changing the version number.